organize new-landing-page
