--- src/components/new-landing-page/astronaut/astronaut-star-group.tsx
@@ -1,33 +1,0 @@
-import React from 'react'
-
-export default function AstronautStarGroup(): JSX.Element {
-  return (
-    <svg viewBox="0 0 529 546" fill="none" className="astronaut-star-group">
-      <path d="M217.5 523.686c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5z" fill="#8BA0BB">
-        <animate attributeName="opacity" values="0.26;0.95;0.26" dur="3.2s" repeatCount="indefinite" />
-      </path>
-      <path d="M190.5 205.5c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5z" fill="#8BA0BB">
-        <animate attributeName="opacity" values="0.36;0.95;0.36" dur="3.1s" repeatCount="indefinite" />
-      </path>
-      <path d="M516.5 471.5c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5z" fill="#DFDFDF">
-        <animate attributeName="opacity" values="0.5;0.95;0.5" dur="3s" repeatCount="indefinite" />
-      </path>
-      <path
-        d="M404.5 539.5c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5zM439.5 140.5c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5zM46.5 153.5c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5z"
-        fill="#8BA0BB"
-        fillOpacity={0.26}
-      />
-      <path d="M181.5 113.5c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5z" fill="#8BA0BB">
-        <animate attributeName="opacity" values="0.18;0.95;0.18" dur="3.3s" repeatCount="indefinite" />
-      </path>
-      <path d="M15.5 5.5c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5z" fill="#8BA0BB">
-        <animate attributeName="opacity" values="0.1;0.95;0.1" dur="3.2s" repeatCount="indefinite" />
-      </path>
-      <path
-        d="M9.5 296.5c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5zM311.5 519.5c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5zM73.5 513.5c-4 2.4-4 4.5-4.5 6-.5-1-1-3.1-5-5.5 3-3 3.5-4 4.5-6 .5 1 1 3.1 5 5.5z"
-        fill="#8BA0BB"
-        fillOpacity={0.26}
-      />
-    </svg>
-  )
-}
