--- src/components/new-landing-page/astronaut/astronaut.tsx
@@ -1,211 +1,0 @@
-import * as React from 'react'
-
-export default () => (
-  <svg viewBox="0 0 293 325" fill="none" className="astronaut">
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M91.115 83.939c-4.114-.58-8.05-2.27-10.703-5.722-7.646-9.946 7.216-19.443 13.941-16.965 6.726 2.478 17.054 14.733 14.526 18.484-1.009 1.495-4.524 3.138-8.872 3.942l12.934 11.006-10.501 9.311-11.325-20.056z"
-      fill="#FFB287"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M85.641 59.782c-1.823.835-4.037 2.534-4.825 3.41a10.1 10.1 0 00-2.852-4.294c-1.273-1.12-2.785-1.693-4.302-2.355-.698-.305-1.427-.579-2.183-.616-.77-.037-1.529.16-2.19.518a5.304 5.304 0 00-1.709 1.514c-.202.272-.324.583-.512.863-.192.285-.451.237-.772.24-.771.01-1.577.39-2.188.812-.686.475-1.246 1.138-1.656 1.865-.4.708-.666 1.52-.666 2.357 0 .373.042.76.166 1.119.05.147.1.218.027.352-.067.125-.182.229-.254.355-.694 1.215-.862 2.732-.452 4.125.24.812 1.008 1.936 2.047 3.037 1.039 1.102 4.68 3.15 6.593 3.38 1.706.205 3.524-.273 4.98-1.069.012 1.632.494 3.254 1.134 4.772a21.152 21.152 0 002.69 4.633c1.037 1.358 2.224 2.715 3.63 3.698 1.315.92 2.762 1.347 4.3 1.069.704-.127 1.382-.355 2.046-.603.655-.244 1.307-.495 1.912-.842 1.144-.658 2.228-1.68 2.29-3.11.023-.516.01-1.127-.095-1.64-.038-.184-.066-.302-.257-.345-.23-.052-.492-.014-.723-.009a5.04 5.04 0 01-3.547-1.315c-1.028-.92-1.925-2.486-1.325-3.817.74-1.643 2.657-2.21 4.274-2.588-.99-2.052-1.345-4.375-.516-6.453.917-2.3 2.73-4.08 3.797-6.302.278-.579.542-1.19.533-1.858-.005-.28-.027-.627-.158-.885-.173-.339-.555-.412-.882-.505-1.736-.493-6.532-.347-8.355.487z"
-      fill="#FA759E"
-    />
-    <path fillRule="evenodd" clipRule="evenodd" d="M122.879 150.515l73.953 41.186 58.919 36.898 6.077-3.508-114.3-88.807-24.649 14.231z" fill="#fff" />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M250.618 228.289c.306 1.839.676 3.135 1.11 3.885.527.914 1.449 2.159 2.765 3.735l28.563-16.491c-.675-2.883-1.868-3.96-3.58-3.232a184.568 184.568 0 01-4.292 1.771l-16.693 1.746a.584.584 0 00-.52.642l.002.017.341 2.563c-1.081 1.545-2.085 2.585-3.01 3.12-.752.434-1.966.73-3.642.887a1.17 1.17 0 00-1.044 1.357z"
-      fill="#FFCA00"
-    />
-    <path
-      d="M282.941 218.82c-.376-.23-19.487 10.703-28.996 16.198-.103.507.376.988.78 1.357.405.37 28.507-15.673 28.775-16.29.269-.617-.089-.978-.559-1.265z"
-      fill="#CBA71E"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M242.998 224.606l9.672-10.311c-13.996-12.384-30.04-25.642-48.134-39.773-18.093-14.13-36.366-27.298-54.817-39.503l-33.098 19.109c22.03 16.195 43.488 30.079 64.373 41.651 20.886 11.572 41.554 21.181 62.004 28.827z"
-      fill="url(#astronaut_paint0_radial)"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M120.406 151.943c9.212 24.158 18.669 56.358 18.526 57.603-.096.83-12.609 14.987-37.541 42.471l4.406 4.6c33.45-26.332 50.319-41.518 50.607-45.556.431-6.058-4.73-48.971-11.243-73.41l-24.755 14.292z"
-      fill="url(#astronaut_paint1_linear)"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M109.472 241.174l7.295 7.296c28.674-13.364 44.399-24.748 47.172-34.154 2.774-9.405-2.844-35.331-16.854-77.776l-31.561 18.221c12.931 33.915 19.478 51.695 19.641 53.338.163 1.643-8.401 12.668-25.693 33.075z"
-      fill="url(#astronaut_paint2_radial)"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M101.6 248.919c-1.758.621-2.97 1.211-3.634 1.768-.807.678-1.873 1.802-3.197 3.371l21.2 25.266c2.722-1.165 3.576-2.527 2.562-4.087a183.73 183.73 0 01-2.49-3.919l-4.618-16.136a.585.585 0 00-.739-.396l-2.465.781c-1.709-.797-2.908-1.604-3.595-2.423-.558-.665-1.06-1.809-1.506-3.432a1.17 1.17 0 00-1.518-.793z"
-      fill="#FFCA00"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M161.672 99.68l9.705-12.674c-1.214-7.627-.582-12.012 1.898-13.156 1.526-1.738 1.988 8.372 5.297 6.02 3.308-2.352 9.28-12.868 11.4-12.201 3.138.987 3.803 7.968 2.759 12.046-1.95 7.622-4.777 11.18-12.757 17.723-4.172 3.421-8.143 8.344-11.914 14.769l-6.388-12.528z"
-      fill="url(#astronaut_paint3_linear)"
-    />
-    <mask id="a" maskUnits="userSpaceOnUse" x={154} y={66} width={50} height={47}>
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M161.672 99.678l9.705-12.673c-1.214-7.627-.582-12.012 1.898-13.156 1.526-1.738 1.988 8.372 5.297 6.02 3.308-2.352 9.28-12.868 11.4-12.201 3.138.987 3.803 7.968 2.759 12.046-1.95 7.622-4.777 11.18-12.757 17.723-4.172 3.421-8.143 8.344-11.914 14.769l-6.388-12.528z"
-        fill="url(#astronaut_paint4_linear)"
-      />
-    </mask>
-    <g mask="url(#a)">
-      <path
-        d="M170.696 84.804c-.406.26-.048 2.79.181 4.022.399.254 5.037 8.732 8.624 9.674 3.587.942 6.087-2.862 6.413-3.623.326-.761-9.565-9.312-11.522-10.4-1.957-1.086-3.189 0-3.696.327z"
-        fill="#E1E1E1"
-      />
-    </g>
-    <path
-      d="M140.783 87.82l-19.472 27.096 9.445 2.409 15.154-25.046c6.425-6.517 9.03-10.075 7.815-10.673-1.823-.898-3.159 1.513-4.487 1.106-1.329-.407-1.541-4.389-1.581-8.176-.04-3.787-1.617-2.544-2.365-1.565-1.962 4.041-3.465 8.991-4.509 14.85z"
-      fill="#E1E1E1"
-    />
-    <mask id="b" maskUnits="userSpaceOnUse" x={109} y={71} width={55} height={50}>
-      <path
-        d="M140.783 87.82l-19.472 27.096 9.445 2.409 15.154-25.046c6.425-6.517 9.03-10.075 7.815-10.673-1.823-.898-3.159 1.513-4.487 1.106-1.329-.407-1.541-4.389-1.581-8.176-.04-3.787-1.617-2.544-2.365-1.565-1.962 4.041-3.465 8.991-4.509 14.85z"
-        fill="#E1E1E1"
-      />
-    </mask>
-    <g mask="url(#b)">
-      <path
-        d="M149.401 89.84c.052-.391-5.763-2.85-8.676-4.03-.341-.358-1.609 2.746-1.723 3.59-.113.846 6.808 3.055 8.287 3.348 1.479.292 2.047-2.421 2.112-2.909z"
-        fill="#B9B9B9"
-      />
-    </g>
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M120.508 106.957c6.119-3.62 13.579-9.66 19.129-20.279l7.87 5.686c-2.581 9.486-8.233 19.319-15.782 23.399-5.95 3.216-13.031-2.138-11.217-8.806z"
-      fill="#E1E1E1"
-    />
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M135.902 114.475c7.478-2.144 17.329-10.395 29.554-24.753 8.374 5.692 13.388 10.302 15.041 13.831-9.218 14.515-18.756 24.12-27.958 29.703 3.489 5.36 5.854 11.01 6.368 17.124 1.059 12.568-31.447 19.13-39.559 26.215-16.478-15.304-9.763-29.267-16.631-41.182-7.52-13.048-8.638-39.17-7.097-40.06l14.051-8.113c6.527 9.514 16.767 18.238 26.231 27.235z"
-      fill="url(#astronaut_paint5_linear)"
-    />
-    <path d="M181 106.186c0-1.2-12.333-11.5-18.5-16.5 0-1.5 1.5-7 3.5-7s20.5 17 21 17.5-6 7.5-6 6z" fill="url(#astronaut_paint6_linear)" />
-    <path
-      d="M147.829 96.465c-.112-.597-7.497-4.516-11.176-6.4-.14-.747.128-3.63 1.17-3.825 1.042-.195 12.265 6.454 12.572 6.654.307.2-2.426 4.317-2.566 3.571z"
-      fill="#E1E1E1"
-    />
-    <rect x={89.7051} y={168.802} width={10.2806} height={4.27451} rx={2} transform="rotate(-5.52 89.705 168.802)" fill="#E1E1E1" />
-    <path
-      d="M76.224 98.938c1.805-.67 3.423 1.349 2.373 2.964l-6.276 9.665a2 2 0 01-1.57.908l-1.255.067a2 2 0 01-2.104-1.89l-.364-6.799a2 2 0 011.3-1.982l7.896-2.933zM114.828 143.58c5.484 5.483 24.645-4.382 28.974-6.708-4.845 4.921-27.611 16.198-30.815 12.816-3.484-3.677-8.998-29.335-8.417-26.819.58 2.516 4.21 14.662 10.258 20.711zM142.202 163.203c.434-.353.832-.661 1.17-.911-.145.197-.557.511-1.17.911-2.349 1.912-5.774 5.138-6.695 7.687-1.363 3.775-10.067 9.438-10.486 9.018-.42-.419-3.461-5.977-2.202-6.501 1.1-.459 15.114-7.414 19.383-10.204zM154.085 158.394c-.214 1.215 4.456 16.669 6.838 25.512.398 1.484 7.681 4.377 6.609 3.93-1.072-.446-13.179-30.96-13.447-29.442zM64.231 118.71c1.367-3.281 7.108-4.269 10.064-4.926 3.26-.724 8.437-.816 9.53.916 1.094 1.732 10.116 49.668 9.387 50.489-.729.82-18.136 2.916-20.323 1.549-2.187-1.367-10.025-44.747-8.658-48.028z"
-      fill="#E1E1E1"
-    />
-    <path
-      d="M70.574 103.315c1.847-4.432 9.601-5.767 13.594-6.654 4.403-.979 11.24-2.957 12.717-.618 1.477 2.339 13.822 68.948 12.837 70.056-.985 1.108-24.498 3.939-27.453 2.093-2.954-1.847-13.541-60.446-11.695-64.877z"
-      fill="#F5F5F5"
-    />
-    <mask id="c" maskUnits="userSpaceOnUse" x={64} y={113} width={30} height={55}>
-      <path
-        d="M64.231 118.71c1.367-3.281 7.108-4.269 10.064-4.926 3.26-.724 8.437-.816 9.53.916 1.094 1.732 10.116 49.668 9.387 50.489-.729.82-18.136 2.916-20.323 1.549-2.187-1.367-10.025-44.747-8.658-48.028z"
-        fill="#fff"
-      />
-    </mask>
-    <g mask="url(#c)">
-      <path d="M70.152 133.356c-.046-.23-3.135-17.745-3.826-18.252" stroke="#FFCA00" strokeWidth={3} />
-    </g>
-    <path
-      d="M86.554 161.419c-1.076-.409-5.442-24.241-7.49-36.105-.384-11.587 14.468-11.843 16.452-3.393.278 1.183 8.194 37.066 8.002 37.962-.192.896-15.62 2.049-16.964 1.536z"
-      fill="#fff"
-    />
-    <path
-      d="M84.537 158.589c-1.267-.483-6.41-28.558-8.823-42.535-.453-13.65 17.044-13.952 19.382-3.997.327 1.393 9.653 43.666 9.427 44.722-.226 1.056-18.402 2.414-19.986 1.81z"
-      fill="url(#astronaut_paint7_linear)"
-    />
-    <path
-      d="M84.881 160.255c-.86-.795-.8-1.711-.662-2.069 1.49-2.07 19.621-3.395 20.366-2.732.745.662.165 1.821 0 2.732-.166.91-18.628 3.063-19.704 2.069z"
-      fill="#E1E1E1"
-    />
-    <path
-      d="M137.1 127.746c4.823 5.719 7.982 7.147 9.44 7.089-3.944 4.17-23.086 12.193-26.047 8.483-2.667-3.341-9.786-21.011-11.724-24.912-.287-.182-.454-.372-.483-.572-.057-.4.125-.149.483.572 3.372 2.147 23.248 3.313 28.331 9.34z"
-      fill="#fff"
-    />
-    <circle cx={90.1172} cy={62} r={34} fill="url(#astronaut_paint8_linear)" />
-    <mask id="d" maskUnits="userSpaceOnUse" x={56} y={28} width={69} height={68}>
-      <circle cx={90.1172} cy={62} r={34} fill="#fff" />
-    </mask>
-    <g mask="url(#d)">
-      <path
-        d="M118.344 80.162c2.219-1.66-7.248 11.363-11.583 14.738 0 .698-18.414 7.767-21.207 6.894-2.793-.872-.698-.785-1.047-3.316-.35-2.53 16.058-9.513 17.891-10.473 1.832-.96 9.795-3.24 15.946-7.843z"
-        fill="#FEFFFF"
-      />
-    </g>
-    <path
-      d="M66.774 47.726c23.37-20.545 43.94-3.386 42.805-5.24-5.109-8.342-14.742-11.653-21.468-10.895-9.607 1.083-18.101 5.474-21.337 16.135zM118.833 57.955c-.417-1.842-2.203-5.487-3.053-6.797 2.741.138 3.29 1.938 3.677 2.473.387.536-.206 6.166-.624 4.324z"
-      fill="#164176"
-    />
-    <circle cx={113.479} cy={108.703} r={7.33816} fill="#FFCA00" />
-    <path d="M117.983 169.716c-1.23-2.013-9.842-14.762-10.625-15.881.586 8.484 6.339 19.48 9.869 20.439 1.08.293 1.986-2.545.756-4.558z" fill="#E1E1E1" />
-    <defs>
-      <radialGradient
-        id="astronaut_paint0_radial"
-        cx={0}
-        cy={0}
-        r={1}
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="rotate(129.581 81.44 124.889) scale(113.176 79.9347)"
-      >
-        <stop stopColor="#C3C3C3" />
-        <stop offset={0.52403} stopColor="#fff" />
-        <stop offset={1} stopColor="#C8C8C8" />
-      </radialGradient>
-      <linearGradient id="astronaut_paint1_linear" x1={103.986} y1={161.423} x2={145.659} y2={233.603} gradientUnits="userSpaceOnUse">
-        <stop stopColor="#fff" />
-        <stop offset={1} stopColor="#F9F9F9" />
-      </linearGradient>
-      <radialGradient
-        id="astronaut_paint2_radial"
-        cx={0}
-        cy={0}
-        r={1}
-        gradientUnits="userSpaceOnUse"
-        gradientTransform="rotate(11.777 -938.893 660.85) scale(48.001 49.107)"
-      >
-        <stop stopColor="#5A5A5A" />
-        <stop offset={0.596197} stopColor="#EEE" />
-        <stop offset={1} stopColor="#fff" />
-      </radialGradient>
-      <linearGradient id="astronaut_paint3_linear" x1={172.249} y1={74.2245} x2={185.58} y2={104.49} gradientUnits="userSpaceOnUse">
-        <stop stopColor="#fff" />
-        <stop offset={1} stopColor="#F9F9F9" />
-      </linearGradient>
-      <linearGradient id="astronaut_paint4_linear" x1={172.249} y1={74.2235} x2={185.58} y2={104.489} gradientUnits="userSpaceOnUse">
-        <stop stopColor="#fff" />
-        <stop offset={1} stopColor="#F9F9F9" />
-      </linearGradient>
-      <linearGradient id="astronaut_paint5_linear" x1={116.175} y1={83.4857} x2={158.522} y2={156.833} gradientUnits="userSpaceOnUse">
-        <stop stopColor="#fff" />
-        <stop offset={1} stopColor="#F9F9F9" />
-      </linearGradient>
-      <linearGradient id="astronaut_paint6_linear" x1={174.764} y1={82.6855} x2={174.764} y2={106.396} gradientUnits="userSpaceOnUse">
-        <stop stopColor="#fff" />
-        <stop offset={1} stopColor="#F9F9F9" />
-      </linearGradient>
-      <linearGradient id="astronaut_paint7_linear" x1={103.724} y1={126.425} x2={77.8117} y2={130.639} gradientUnits="userSpaceOnUse">
-        <stop stopColor="#EFEFEF" />
-        <stop offset={0.534453} stopColor="#fff" stopOpacity={0.465547} />
-        <stop offset={1} stopColor="#EFEFEF" stopOpacity={0} />
-      </linearGradient>
-      <linearGradient id="astronaut_paint8_linear" x1={105.71} y1={8.80325} x2={90.1172} y2={96} gradientUnits="userSpaceOnUse">
-        <stop stopColor="#FA759E" stopOpacity={0.77} />
-        <stop offset={1} stopColor="#164176" stopOpacity={0.12} />
-      </linearGradient>
-    </defs>
-  </svg>
-)
