--- src/components/new-landing-page/donate.tsx
@@ -1,44 +1,0 @@
-import React from 'react'
-
-function BuyMeACoffee(): JSX.Element {
-  return (
-    <svg viewBox="0 0 884 1279" fill="none" className="buymeacoffee">
-      <path
-        d="M791.109 297.518l-.878-.516-2.03-.619a4.833 4.833 0 002.908 1.135zM803.896 388.891l-.98.275.98-.275zM791.484 297.377a1.773 1.773 0 01-.366-.087 1.42 1.42 0 000 .244.738.738 0 00.366-.157z"
-        fill="#0D0C22"
-      />
-      <path
-        d="M791.113 297.529h.131v-.082l-.131.082zM803.111 388.726l1.48-.843.551-.31.499-.533a8.447 8.447 0 00-2.53 1.686zM793.669 299.515l-1.446-1.377-.98-.533a4.077 4.077 0 002.426 1.91zM430.019 1186.18a7.55 7.55 0 00-2.943 2.27l.912-.58c.62-.57 1.497-1.24 2.031-1.69zM641.187 1144.63c0-1.3-.636-1.06-.482 3.58 0-.37.155-.75.224-1.11.086-.83.155-1.64.258-2.47zM619.284 1186.18a7.546 7.546 0 00-2.942 2.27l.912-.58c.619-.57 1.497-1.24 2.03-1.69zM281.304 1196.06a6.299 6.299 0 00-3.097-1.45c.929.45 1.858.9 2.477 1.24l.62.21zM247.841 1164.01a9.895 9.895 0 00-1.222-3.85c.474 1.23.87 2.5 1.187 3.78l.035.07z"
-        fill="#0D0C22"
-      />
-      <path
-        className="logo-coffee"
-        d="M472.623 590.836c-45.941 19.667-98.077 41.966-165.647 41.966a313.577 313.577 0 01-83.623-11.528l46.733 479.806a80.166 80.166 0 0025.593 52.38 80.18 80.18 0 0054.313 21.19s66.262 3.44 88.373 3.44c23.796 0 95.151-3.44 95.151-3.44 20.12 0 39.503-7.57 54.303-21.2a80.149 80.149 0 0025.587-52.37l50.053-530.204c-22.368-7.639-44.943-12.715-70.391-12.715-44.014-.017-79.477 15.142-120.445 32.675z"
-        fill="#FD0"
-      />
-      <path d="M78.689 386.132l.79.74.517.31a7.95 7.95 0 00-1.307-1.05z" fill="#0D0C22" />
-      <path
-        className="logo-outline"
-        d="M879.567 341.849l-7.037-35.497c-6.315-31.849-20.648-61.943-53.34-73.454-10.479-3.683-22.369-5.265-30.404-12.888-8.035-7.622-10.41-19.46-12.268-30.438-3.442-20.149-6.676-40.315-10.204-60.429-3.045-17.293-5.454-36.719-13.386-52.583-10.324-21.302-31.746-33.76-53.048-42.001a305.493 305.493 0 00-33.363-10.324C613.297 10.195 557.342 5.033 502.591 2.09a1376.266 1376.266 0 00-197.169 3.27c-48.797 4.439-100.193 9.807-146.564 26.687-16.948 6.177-34.413 13.593-47.3 26.687-15.813 16.088-20.975 40.969-9.43 61.031 8.208 14.247 22.111 24.313 36.857 30.972a298.948 298.948 0 0059.844 19.478c57.297 12.664 116.642 17.636 175.178 19.753a1333.99 1333.99 0 00194.433-6.35 1106.995 1106.995 0 0047.817-6.314c18.738-2.874 30.765-27.376 25.242-44.445-6.608-20.406-24.365-28.321-44.444-25.241-2.96.464-5.902.894-8.862 1.324l-2.133.31c-6.803.861-13.605 1.663-20.407 2.409a1083.055 1083.055 0 01-42.259 3.717c-31.626 2.202-63.337 3.217-95.031 3.269-31.144 0-62.305-.878-93.38-2.925a1187.382 1187.382 0 01-42.431-3.545 1019.94 1019.94 0 01-19.219-2.168l-6.092-.774-1.324-.189-6.315-.912c-12.905-1.945-25.81-4.181-38.577-6.883a5.8 5.8 0 010-11.322h.241c11.064-2.357 22.213-4.37 33.397-6.125 3.729-.585 7.468-1.159 11.219-1.72h.103c7.003-.465 14.041-1.722 21.009-2.547A1336.167 1336.167 0 01469.538 73.1c29.577.86 59.138 2.599 88.578 5.593a979.44 979.44 0 0118.927 2.116c2.409.293 4.835.636 7.262.93l4.886.705a678.058 678.058 0 0142.517 7.725c20.889 4.543 47.714 6.022 57.005 28.907 2.96 7.261 4.302 15.331 5.936 22.953l2.083 9.722c.054.174.094.353.12.533 4.921 22.942 9.848 45.884 14.78 68.826a12.608 12.608 0 01-2.006 9.871 12.612 12.612 0 01-8.593 5.254h-.138l-3.011.413-2.976.395c-9.43 1.228-18.87 2.375-28.322 3.442a1829.308 1829.308 0 01-55.938 5.506 1957.204 1957.204 0 01-111.55 6.074c-18.984.504-37.963.74-56.936.705a1975.736 1975.736 0 01-225.989-13.146c-8.122-.963-16.243-1.996-24.365-3.045 6.298.809-4.577-.62-6.779-.929a1447.245 1447.245 0 01-15.486-2.254c-17.327-2.599-34.55-5.799-51.843-8.604-20.906-3.441-40.9-1.72-59.81 8.604a86.994 86.994 0 00-36.012 37.338c-8.156 16.862-10.582 35.221-14.23 53.34C4 342.193-1.678 361.688.473 380.288 5.1 420.431 33.165 453.054 73.53 460.35c37.975 6.882 76.156 12.457 114.44 17.206a2114.804 2114.804 0 00489.988 2.822 25.814 25.814 0 0121.003 7.339 25.806 25.806 0 017.491 20.948l-3.82 37.132-23.091 225.077a178840.62 178840.62 0 01-31.126 302.866c-2.203 21.84-2.512 44.36-6.659 65.94-6.539 33.93-29.509 54.77-63.027 62.39a439.172 439.172 0 01-93.569 10.94c-34.912.19-69.806-1.36-104.718-1.17-37.27.21-82.918-3.23-111.687-30.97-25.277-24.36-28.77-62.51-32.211-95.5-4.588-43.67-9.136-87.331-13.645-130.989l-25.293-242.766-16.363-157.077c-.276-2.598-.551-5.162-.809-7.778-1.962-18.737-15.228-37.079-36.134-36.133-17.894.791-38.232 16.002-36.133 36.133l12.13 116.454 25.087 240.89a378681.11 378681.11 0 0121.388 205.306c1.377 13.11 2.667 26.26 4.112 39.37 7.864 71.65 62.58 110.26 130.339 121.13 39.575 6.37 80.113 7.68 120.273 8.33 51.482.83 103.48 2.81 154.118-6.52 75.038-13.77 131.337-63.87 139.372-141.59 2.295-22.44 4.589-44.88 6.883-67.33 7.628-74.241 15.245-148.487 22.85-222.739l24.881-242.61 11.408-111.188a25.812 25.812 0 0120.785-22.696c21.456-4.181 41.967-11.322 57.229-27.651 24.295-25.998 29.13-59.895 20.544-94.067zM72.43 365.835c.327-.155-.275 2.649-.533 3.957-.052-1.979.051-3.734.533-3.957zm2.082 16.105c.172-.121.688.568 1.222 1.394-.809-.758-1.325-1.325-1.24-1.394h.018zm2.048 2.701c.74 1.256 1.135 2.048 0 0zm4.112 3.338h.103c0 .121.19.241.258.362a2.666 2.666 0 00-.378-.362h.017zm720.124-4.99c-7.708 7.33-19.323 10.737-30.8 12.441-128.704 19.099-259.283 28.769-389.399 24.502-93.121-3.183-185.261-13.525-277.453-26.55-9.034-1.273-18.824-2.925-25.036-9.584-11.7-12.561-5.953-37.854-2.908-53.03 2.788-13.903 8.122-32.434 24.657-34.413 25.81-3.028 55.783 7.863 81.318 11.735a1539.798 1539.798 0 0092.57 11.27c132.18 12.045 266.58 10.169 398.175-7.45a1661.103 1661.103 0 0071.699-11.236c21.216-3.803 44.737-10.943 57.556 11.029 8.792 14.97 9.962 34.998 8.603 51.912a28.942 28.942 0 01-8.999 19.374h.017z"
-        fill="#0D0C22"
-      />
-    </svg>
-  )
-}
-
-export default ({ makeAndTrackRef }: { makeAndTrackRef(): React.MutableRefObject<any> }) => (
-  <div className="donate-container">
-    <div className="donate">
-      <h1 className="mhi-heading human-blue" ref={makeAndTrackRef()}>
-        Support our work
-      </h1>
-      <p className="human-blue" ref={makeAndTrackRef()}>
-        Your donation will help keep us and our servers running
-      </p>
-      <a ref={makeAndTrackRef()} className="donate-link mhi-button umami--click--donate" href="https://buymeacoffee.com/morehumaninter">
-        <BuyMeACoffee />
-        Buy us a coffee
-      </a>
-    </div>
-  </div>
-)
