--- src/components/new-landing-page/mountains/mountain-background.tsx
@@ -1,31 +1,0 @@
-import * as React from 'react'
-
-export function MountainBackground1(): JSX.Element {
-  return (
-    <svg viewBox="0 0 359 258" fill="none" className="mountain-background1" preserveAspectRatio="none">
-      <path
-        d="M231.116.336L153 31-.438 105.726a1.001 1.001 0 00-.562.899V257a1 1 0 001 1h357.393a1 1 0 00.897-1.442L232.379.825a1 1 0 00-1.263-.489z"
-        fill="#FFA0BF"
-      />
-    </svg>
-  )
-}
-
-export function MountainBackground2(): JSX.Element {
-  return (
-    <svg viewBox="0 0 1440 308" fill="none" className="mountain-background2" preserveAspectRatio="none">
-      <path
-        d="M563.07 29.146l-99.109 61.678h-.001c-.141.088-.211.132-.287.163-.077.032-.158.05-.319.088L1.55 197.642c-.743.172-1.115.258-1.333.531-.217.273-.217.655-.217 1.418V306c0 .943 0 1.414.293 1.707C.586 308 1.057 308 2 308h1436c.94 0 1.41 0 1.71-.293.29-.293.29-.764.29-1.707V107.522c0-.259 0-.389-.03-.514-.03-.124-.1-.237-.23-.464l-16.94-30.24c-.59-1.057-.89-1.585-1.38-1.654-.5-.068-.92.361-1.78 1.218l-25.76 25.765c-.18.183-.27.274-.38.339-.11.064-.23.1-.48.17l-38.16 10.907c-1.66.475-2.49.712-3.35.803-.85.092-1.71.037-3.43-.074L1242.57 107 1119.9 91.037c-.15-.018-.22-.028-.29-.027-.07.001-.14.013-.28.035l-68.5 10.866c-.28.044-.42.066-.55.05-.14-.017-.27-.072-.53-.181L911.647 43.827c-.384-.161-.577-.242-.779-.241-.203 0-.395.082-.779.246L800.371 90.605c-.454.194-.682.29-.919.276-.237-.015-.45-.14-.876-.39l-60.12-35.217c-.233-.136-.349-.204-.478-.24-.129-.034-.264-.034-.533-.034h-26.766c-.333 0-.5 0-.655-.052-.155-.052-.287-.153-.553-.354L639.358 1.485c-.623-.471-.935-.707-1.291-.694-.356.012-.649.27-1.236.786l-46.643 40.96c-.529.464-.794.697-1.119.726-.324.03-.626-.152-1.23-.515L565.156 29.13c-.51-.306-.765-.46-1.047-.457-.281.002-.534.16-1.039.474z"
-        fill="url(#mountain-background2_paint0_linear)"
-      />
-      <defs>
-        <linearGradient id="mountain-background2_paint0_linear" x1={1403.03} y1={-131.499} x2={582.964} y2={163.138} gradientUnits="userSpaceOnUse">
-          <stop stopColor="#B46692" />
-          <stop offset={0.491021} stopColor="#7C5988" />
-          <stop offset={0.732053} stopColor="#E4709B" />
-          <stop offset={1} stopColor="#985F8D" />
-        </linearGradient>
-      </defs>
-    </svg>
-  )
-}
