--- src/components/new-landing-page/mountains/mountain-midground.tsx
@@ -1,18 +1,0 @@
-import * as React from 'react'
-
-export default function MountainMidground(props: any): JSX.Element {
-  return (
-    <svg viewBox="0 0 1440 303" fill="none" className="mountain-midground" preserveAspectRatio="none" {...props}>
-      <path
-        d="M138.172.257L66.432 25.5 1.165 55.465c-.567.26-.85.39-1.008.636C0 56.348 0 56.66 0 57.282V301c0 .943 0 1.414.293 1.707C.586 303 1.057 303 2 303h1436c.94 0 1.41 0 1.71-.293.29-.293.29-.764.29-1.707V159.919c0-.452 0-.678-.09-.877-.09-.199-.26-.346-.61-.641l-72.47-62.185c-.13-.108-.19-.161-.26-.204-.07-.043-.15-.074-.3-.135l-103.36-41.263-.09-.037c-.09-.033-.14-.053-.2-.066-.07-.017-.15-.023-.3-.034l-144.38-10.96c-.11-.009-.16-.013-.22-.011-.05.002-.1.009-.21.024L924.537 70.5 878.391 82l-41.83 9.403c-.215.048-.322.072-.432.073-.109 0-.217-.023-.432-.07l-161.773-35.27c-.31-.067-.465-.101-.62-.086-.156.015-.301.077-.593.202l-81.652 34.96c-.333.142-.5.214-.677.223-.177.01-.351-.044-.697-.15L414.982 37.627c-.206-.063-.309-.095-.415-.104-.107-.01-.214.005-.427.032L271.257 55.931c-.268.034-.402.051-.534.033-.131-.018-.256-.07-.504-.176L139.618.303c-.353-.15-.529-.225-.716-.23-.187-.006-.368.057-.73.184z"
-        fill="url(#mountain-midground_paint0_linear)"
-      />
-      <defs>
-        <linearGradient id="mountain-midground_paint0_linear" x1={904.06} y1={-321.999} x2={716.157} y2={211.878} gradientUnits="userSpaceOnUse">
-          <stop stopColor="#C36995" />
-          <stop offset={1} stopColor="#174277" />
-        </linearGradient>
-      </defs>
-    </svg>
-  )
-}
