+++ src/styles/components/header.scss
@@ -94,176 +94,6 @@ svg.header-logo {
   }
 }
 
-@keyframes rotate-nothing {
-  0% {
-    opacity: 1;
-  }
-
-  100% {
-    opacity: 1;
-  }
-}
-
-@keyframes rotate-first {
-  0% {
-    opacity: 1;
-    transform: translate3d(0, 0, 0);
-  }
-
-  20%,
-  80% {
-    opacity: 1;
-    transform: translate3d(0, 0, 0);
-  }
-
-  100% {
-    opacity: 0;
-    transform: translate3d(-25px, 0, 0);
-  }
-}
-
-@keyframes rotate {
-  0% {
-    opacity: 0;
-    transform: translate3d(50px, 0, 0);
-  }
-
-  20%,
-  80% {
-    opacity: 1;
-    transform: translate3d(0, 0, 0);
-  }
-
-  100% {
-    opacity: 0;
-    transform: translate3d(-25px, 0, 0);
-  }
-}
-
-@keyframes rotate-last {
-  0% {
-    opacity: 0;
-    transform: translate3d(50px, 0, 0);
-  }
-
-  50%,
-  100% {
-    opacity: 1;
-    transform: translate3d(0, 0, 0);
-  }
-}
-
-@keyframes dots-in {
-  0% {
-    opacity: 0;
-    transform: translate3d(100px, 0, 0);
-  }
-
-  50% {
-    opacity: 0;
-  }
-
-  100% {
-    opacity: 1;
-    transform: translate3d(0, 0, 0);
-  }
-}
-
-@keyframes words-right {
-  100% {
-    transform: translate3d(0, 0, 0);
-  }
-}
-
-@keyframes buttons-down {
-  0% {
-    opacity: 0;
-    transform: translate3d(0, -15px, 0);
-  }
-
-  50% {
-    opacity: 0;
-  }
-
-  100% {
-    opacity: 1;
-    transform: translate3d(0, +50px, 0);
-  }
-}
-
-.mhi-rotating-logo {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.mhi-rotating-logo-dots {
-  position: relative;
-  height: 120px;
-  opacity: 0;
-
-  .bottom-left {
-    fill: $human-blue;
-  }
-  .top-right {
-    fill: $human-pink;
-  }
-
-  animation-name: dots-in;
-  animation-fill-mode: forwards;
-}
-
-.mhi-rotating-logo-words {
-  animation-name: words-right;
-  animation-duration: 1.5s;
-  animation-fill-mode: forwards;
-  transform: translate3d(-50px, 0, 0);
-}
-
-.rotatingText {
-  font-style: italic;
-  position: relative;
-}
-
-.rotatingText-adjective {
-  font-family: "Open Sans", sans-serif;
-  font-size: 50px;
-  font-style: normal;
-  font-weight: 700;
-  left: 0;
-  margin-bottom: 0;
-  opacity: 0;
-  position: absolute;
-  right: 0;
-  text-align: left;
-  text-transform: uppercase;
-  top: 0;
-  line-height: 42px;
-  animation-duration: 1.5s;
-  animation-name: rotate;
-  display: block;
-  white-space: nowrap;
-}
-
-.mhi-rotating-buttons {
-  padding: 1rem;
-  .rotatingButton-adjective {
-    width: 310px;
-    height: 70px;
-    opacity: 0;
-    margin: 0 0.5rem 0 0.5rem;
-    animation-name: buttons-down;
-    animation-fill-mode: forwards;
-
-    &:nth-of-type(1) {
-      background-color: $human-blue;
-    }
-    &:nth-of-type(2) {
-      background-color: $human-pink;
-    }
-  }
-}
-
 header.layout-new-header {
   height: $header-height;
   position: fixed;
